refactor: cleanup sidebar component
