+++ src/components/sidebar/defaultSidebar.tsx
@@ -1,0 +1,89 @@
+import {Fragment, SVGAttributes} from 'react';
+import {Link} from 'react-feather';
+
+import styles from './style.module.scss';
+
+import {DefaultSidebarProps, SidebarNode} from './types';
+
+export function DefaultSidebar({node, path}: DefaultSidebarProps) {
+  const activeClassName = (n: SidebarNode, baseClassName = '') => {
+    const className = n.path === path.join('/') ? 'active' : '';
+    return `${baseClassName} ${className}`;
+  };
+
+  const renderChildren = (children: SidebarNode[]) =>
+    children && (
+      <ul data-sidebar-tree>
+        {children
+          .filter(n => n.frontmatter.sidebar_title || n.frontmatter.title)
+          .map(n => (
+            <li className="toc-item" key={n.path} data-sidebar-branch>
+              <Link
+                href={'/' + n.path}
+                data-sidebar-link
+                className={activeClassName(n, 'flex items-center justify-between gap-1')}
+              >
+                {n.frontmatter.sidebar_title || n.frontmatter.title}
+                {n.children.length > 0 && <Chevron direction="down" />}
+              </Link>
+              {renderChildren(n.children)}
+            </li>
+          ))}
+      </ul>
+    );
+
+  const rotation = {
+    down: 0,
+    right: 270,
+  } as const;
+
+  function Chevron({
+    direction,
+    ...props
+  }: SVGAttributes<SVGElement> & {
+    direction: 'down' | 'right';
+  }) {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        viewBox="0 0 16 16"
+        {...props}
+        style={{
+          transition: 'transform 200ms',
+          transform: `rotate(${rotation[direction]}deg)`,
+        }}
+      >
+        <path
+          fill="currentColor"
+          d="M12.53 5.47a.75.75 0 0 1 0 1.06l-4 4a.75.75 0 0 1-1.06 0l-4-4a.75.75 0 0 1 1.06-1.06L8 8.94l3.47-3.47a.75.75 0 0 1 1.06 0Z"
+        />
+      </svg>
+    );
+  }
+
+  return (
+    <ul data-sidebar-tree>
+      <li className="mb-3" data-sidebar-branch>
+        <Fragment>
+          <Link
+            href={'/' + node.path}
+            data-active={node.path === path.join('/')}
+            className={activeClassName(
+              node,
+              [styles['sidebar-title'], 'flex items-center justify-between gap-1'].join(
+                ' '
+              )
+            )}
+            data-sidebar-link
+            key={node.path}
+          >
+            <h6>{node.frontmatter.sidebar_title || node.frontmatter.title}</h6>
+          </Link>
+          {renderChildren(node.children)}
+        </Fragment>
+      </li>
+    </ul>
+  );
+}
