+++ src/components/sidebar/utils.ts
@@ -1,0 +1,30 @@
+import {DocNode} from 'sentry-docs/docTree';
+
+import {NavNode} from './types';
+
+export function getNavNodes<NavNode_>(
+  docNodes: DocNode[],
+  docNodeToNavNode_: (doc: DocNode) => NavNode_ | undefined,
+  nodes: NavNode_[] = []
+) {
+  docNodes.forEach(n => {
+    const navNode = docNodeToNavNode_(n);
+    if (!navNode) {
+      return;
+    }
+    nodes.push(navNode);
+    getNavNodes(n.children, docNodeToNavNode_, nodes);
+  });
+  return nodes;
+}
+
+export const docNodeToNavNode = (node: DocNode): NavNode => ({
+  context: {
+    draft: Boolean(node.frontmatter.draft),
+    title: node.frontmatter.title,
+    sidebar_order: node.frontmatter.sidebar_order,
+    sidebar_title: node.frontmatter.sidebar_title,
+    sidebar_hidden: node.frontmatter.sidebar_hidden,
+  },
+  path: '/' + node.path + '/',
+});
